pyproject.toml
[project]
name = "llm-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-google-genai>=2.1.0",
    "langgraph>=0.3.11",
    "python-dotenv>=1.0.1",
]
